upgrade spring-ai SDK to version 1.0.0-M3 (#11)
and related code changes

+++ src/main/java/org/springframework/samples/petclinic/genai/ChatOptionsProperties.java
@@ -28,6 +28,6 @@ public class ChatOptionsProperties {
 
 	String deploymentName;
 
-	Float temperature;
+	Double temperature;
 
 }
